crud app added, adjusted css on other apps

+++ js/tipCalculator.js
@@ -3,9 +3,9 @@ const calculateTip = () => {
     let tip = document.getElementById("tip").value;
     let numberOfPeople = document.getElementById("numberOfPeople").value;
     if(cost.length && tip.length && numberOfPeople.length){
-        cost = parseInt(cost);
-        tip = parseInt(tip);
-        numberOfPeople = parseInt(numberOfPeople);
+        cost = parseFloat(cost);
+        tip = parseFloat(tip);
+        numberOfPeople = parseFloat(numberOfPeople);
 
         const payPerPerson = (cost + cost * (tip / 100)) / numberOfPeople;
         document.getElementById("pay").innerText = "$" + payPerPerson.toFixed(2);
@@ -15,3 +15,4 @@ const calculateTip = () => {
 // improvements
 // prevent the letting e from being input
 // division by zero
+// if user enters a decimal for percentage change tip formula
